# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paper Company",
    "version": "1.0.0",
    "description": "Witaj w repozytorium Systemu ERP dla Papier Company - zainspirowanego przez kultowy serial \"The Office\". Nasz system jest zaprojektowany, aby wspierać i optymalizować kluczowe procesy biznesowe w firmie zajmującej się dystrybucją papieru, od zarządzania zapasami po finanse i księgowość."
  },
  "paths": {
    "/forgot": {
      "post": {
        "summary": "Wysyła email resetujący hasło.",
        "description": "Endpoint do wysyłania emaila z linkiem resetującym hasło użytkownika.",
        "tags": [
          "Forgot Password"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adres email użytkownika."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomyślnie wysłano email resetujący hasło.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat potwierdzający wysłanie emaila."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Pobiera listę wszystkich użytkowników.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usersList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/user/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Pobiera informacje o użytkowniku na podstawie jego identyfikatora.",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}